requirements.txt
google-genai>=1.5.0
